+++ CHANGELOG.md
@@ -7,6 +7,17 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- App: Add `resource.subscription.toggle` action.
+- App: Add `resource.subscription.status` action.
+- App: Add `resource.subscription.unsubscribe` action.
+- Server: Add option to toggle subscriptions.
+
+### Changed
+
+- App: Rename `resource.subscribe` action to `resource.subscription.subscribe`.
+
 ### Fixed
 
 - App: Fix invalid URLs when uploading assets.